+++ android/app/src/main/java/org/getlantern/mobilesdk/util/DnsDetector.java
@@ -4,9 +4,9 @@ import android.content.Context;
 import android.net.ConnectivityManager;
 import android.net.LinkProperties;
 import android.net.Network;
+import android.net.NetworkCapabilities;
 import android.net.NetworkInfo;
 import android.net.NetworkRequest;
-import android.net.RouteInfo;
 
 import androidx.annotation.NonNull;
 
@@ -16,8 +16,10 @@ import java.net.Inet6Address;
 import java.net.InetAddress;
 import java.net.NetworkInterface;
 import java.net.SocketException;
-import java.net.UnknownHostException;
-import java.util.concurrent.atomic.AtomicReference;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 import static android.content.Context.CONNECTIVITY_SERVICE;
 
@@ -30,9 +32,16 @@ import static android.content.Context.CONNECTIVITY_SERVICE;
 public class DnsDetector {
 
     private static final String TAG = "DnsServersDetector";
+    private static final String DEFAULT_DNS_SERVER = "8.8.8.8";
+    private static final int[] NETWORK_PRIORITY = new int[]{
+            ConnectivityManager.TYPE_ETHERNET,
+            ConnectivityManager.TYPE_WIFI,
+            ConnectivityManager.TYPE_MOBILE
+    };
+
     private final String fakeDnsIP;
-    private final AtomicReference<String> dnsServer = new AtomicReference<>("8.8.8.8");
     private final ConnectivityManager connectivityManager;
+    private final Map<Network, Object> allNetworks = new ConcurrentHashMap<>();
 
     /**
      * Constructor
@@ -40,85 +49,95 @@ public class DnsDetector {
     public DnsDetector(Context context, String fakeDnsIP) {
         this.fakeDnsIP = fakeDnsIP;
         connectivityManager = (ConnectivityManager) context.getSystemService(CONNECTIVITY_SERVICE);
-        update();
         connectivityManager.registerNetworkCallback(
-                new NetworkRequest.Builder().build(),
+                new NetworkRequest.Builder()
+                        .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+                        .addCapability(NetworkCapabilities.NET_CAPABILITY_NOT_VPN)
+                        .build(),
                 new ConnectivityManager.NetworkCallback() {
                     @Override
                     public void onAvailable(@NonNull Network network) {
-                        update();
+                        Logger.debug(TAG, "Adding available network");
+                        allNetworks.put(network, "");
+                    }
+
+                    @Override
+                    public void onLost(@NonNull Network network) {
+                        Logger.debug(TAG, "Removing lost network");
+                        allNetworks.remove(network);
                     }
                 }
         );
     }
 
     public String getDnsServer() {
-        return dnsServer.get();
+        String dnsServer = doGetDnsServer();
+        Logger.debug(TAG, "Using DNS server " + dnsServer);
+        return dnsServer;
     }
 
-    private void update() {
-        // This code only works on LOLLIPOP and higher
-        try {
-            // First pass looks for default route
-            if (!doUpdate(true)) {
-                doUpdate(false);
-            }
-        } catch (Exception ex) {
-            Logger.debug(TAG, "Exception detecting DNS servers using ConnectivityManager method", ex);
+    private String doGetDnsServer() {
+        Network network = findActiveNetwork();
+        if (network == null) {
+            return DEFAULT_DNS_SERVER;
         }
-    }
 
-    private boolean doUpdate(boolean lookForDefault) {
-        for (Network network : connectivityManager.getAllNetworks()) {
-            NetworkInfo networkInfo = connectivityManager.getNetworkInfo(network);
-            if (networkInfo.isConnected()) {
-                LinkProperties linkProperties = connectivityManager.getLinkProperties(network);
-                if (lookForDefault && !linkPropertiesHasDefaultRoute(linkProperties)) {
-                    continue;
-                }
-
-                for (InetAddress element : linkProperties.getDnsServers()) {
-                    String dnsHost = element.getHostAddress();
-                    if (!fakeDnsIP.equals(dnsHost)) {
-                        try {
-                            InetAddress address = network.getByName(dnsHost);
-                            String ip = address.getHostAddress();
-                            if (address instanceof Inet6Address) {
-                                // For IPv6, the DNS server address can be a link-local address.
-                                // For Go to know how to route this, it needs to know the zone
-                                // (interface ID), so we append it to the address here.
-                                NetworkInterface intf = NetworkInterface.getByInetAddress(
-                                        linkProperties.getLinkAddresses().get(0).getAddress());
-                                ip = ip + "%" + intf.getIndex();
-                            }
-                            Logger.debug(TAG, "Setting DNS server to " + ip);
-                            dnsServer.set(ip);
-                            return true;
-                        } catch (UnknownHostException uhe) {
-                            Logger.debug(TAG, "Unable to resolve hostname", uhe);
-                        } catch (SocketException se) {
-                            Logger.debug(TAG, "Unable to get NetworkInterface", se);
+        LinkProperties linkProperties = connectivityManager.getLinkProperties(network);
+        for (InetAddress address : linkProperties.getDnsServers()) {
+            String ip = address.getHostAddress();
+            if (!fakeDnsIP.equals(ip)) {
+                try {
+                    if (address instanceof Inet6Address) {
+                        Inet6Address ipv6Address = (Inet6Address) address;
+                        if (ipv6Address.isLinkLocalAddress()) {
+                            // For IPv6, the DNS server address can be a link-local address.
+                            // For Go to know how to route this, it needs to know the zone
+                            // (interface ID). In some cases, that's missing, in other cases it's a
+                            // name rather than an interface ID (which our Go code can't handle).
+                            // This fixes that.
+                            NetworkInterface intf = NetworkInterface.getByInetAddress(
+                                    linkProperties.getLinkAddresses().get(0).getAddress());
+                            ip = ip.split("%")[0] + "%" + intf.getIndex();
                         }
                     }
+                    return ip;
+                } catch (SocketException se) {
+                    Logger.debug(TAG, "Unable to get NetworkInterface", se);
                 }
             }
         }
 
-        return false;
+        return DEFAULT_DNS_SERVER;
     }
 
-    /**
-     * Returns true if the specified link properties have any default route
-     *
-     * @param linkProperties
-     * @return true if the specified link properties have default route or false otherwise
-     */
-    private boolean linkPropertiesHasDefaultRoute(LinkProperties linkProperties) {
-        for (RouteInfo route : linkProperties.getRoutes()) {
-            if (route.isDefaultRoute()) {
-                return true;
+    private Network findActiveNetwork() {
+        List<Network> networks = new ArrayList(allNetworks.keySet());
+        List<NetworkInfo> networkInfos = new ArrayList<>();
+        for (Network network : networks) {
+            networkInfos.add(connectivityManager.getNetworkInfo(network));
+        }
+
+        // search order goes ETHERNET -> WIFI -> MOBILE
+        for (int type : NETWORK_PRIORITY) {
+            Network activeNetwork = availableNetworkOfType(
+                    networks,
+                    networkInfos,
+                    type);
+            if (activeNetwork != null) {
+                return activeNetwork;
             }
         }
-        return false;
+
+        return null;
+    }
+
+    private Network availableNetworkOfType(List<Network> networks, List<NetworkInfo> networkInfos, int type) {
+        for (int i = 0; i < networkInfos.size(); i++) {
+            NetworkInfo info = networkInfos.get(i);
+            if (info != null && info.isAvailable() && info.getType() == type) {
+                return networks.get(i);
+            }
+        }
+        return null;
     }
 }